OS X modifications for downloader daemon #1173
It uses py2app to build the downloader, which may not be the best option. I'll let Luc look it over and decide if there's a better way


git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@1238 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f2505e558d56851e91ada25758201f3d0ba7eed4

+++ tv/platform/osx/frontend.py
@@ -27,6 +27,7 @@ import time
 import math
 import struct
 import string
+import signal
 import urlparse
 import threading
 
@@ -950,10 +951,16 @@ class UIBackendDelegate:
 
     def launchDownloadDaemon(self, oldpid):
         # Use UNIX style kill
-        os.kill(oldpid, signal.SIGTERM)
-        sleep(1)
-        os.kill(oldpid, signal.SIGKILL)
-        print "WARNING: launchDownloadDaemon not implemented"
+        if oldpid is not None:
+            try:
+                os.kill(oldpid, signal.SIGTERM)
+                sleep(1)
+                os.kill(oldpid, signal.SIGKILL)
+            except:
+                pass
+        p = os.path.normpath(resource.path("../Democracy_Downloader.app"))
+        print p
+        NSWorkspace.sharedWorkspace().launchApplication_(p)
 
 class ExceptionReporterController (NibClassBuilder.AutoBaseClass):
